requirements.txt
appdirs==1.4.3
asgiref==3.2.7
distlib==0.3.0
dj-database-url==0.5.0
Django==2.1.11
django-bootstrap-modal-forms==1.4.1
django-widget-tweaks==1.4.2
django-datetimepicker==3.14
filelock==3.0.12
gunicorn==20.0.4
importlib-metadata==1.6.0
psycopg2==2.8.4
psycopg2-binary==2.8.4
pytz==2018.5
selenium==3.14.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
virtualenv==20.0.15
virtualenvwrapper-win==1.2.6
whitenoise==5.0.1
zipp==3.1.0
django-bootstrap-breadcrumbs==0.9.2
django-currentuser==0.5.0
